move GameState
