+++ src/slashchain/game.ts
@@ -3,7 +3,13 @@ import { INVALID_MOVE } from "boardgame.io/core";
 import { Board } from "./board";
 import { HandsSet } from "./rules";
 import { TileCell, Tile } from "./components";
-import { GameState } from "./view/game_state";
+import { PlayerHands } from "./hands";
+
+export type GameState = {
+  ruleName: string;
+  board: Board;
+  hands: { [key: string]: PlayerHands };
+};
 
 export const Slashchain: Game<GameState> = {
   name: "slashchain",
